Added logic to prevent writing to the dictionary if efficiencyBoost is set to false
Added logic to prevent writing to the dictionary if efficiencyBoost is set to false ⚔️ (Initially, only reading from the dictionary was prevented, for efficiencyBoost: false.  So, it was writing to the dictionary without using the stored values.)  However, this did not seem to significantly impact the speed comparisons & conclusions drawn between execution time of this program vs. the Scheme implementation (as discussed in the note at-bottom in the playground).